+++ NEWS.rst
@@ -4,7 +4,8 @@ Version 0.10.0 (In development)
 New features
 ------------
 
-* #22849: Updated SAML implementation, with support for signed requests and SLO
+* #22849: Updated SAML implementation, with support for signed requests,
+  single sign-on and single logout.
 * #22381: Replace 'Enhedsnummer' with a description of the location of the organisational unit.
 * #23558: Added the possibility to create managers without employees through the ou endpoint, thus allowing for vacant manager positions.